Moved results from App page to Search Page, changed Home and Search functions to classes.
